Remove `air_workspace` for now for simplicity
LSP needs `format_node()` since it holds onto the AST, so this isn't as useful yet over there
