--- crates/air_workspace/Cargo.toml
@@ -1,27 +1,0 @@
-
-[package]
-authors.workspace = true
-categories.workspace = true
-description = "Air workspace utilities"
-edition.workspace = true
-homepage.workspace = true
-keywords.workspace = true
-license.workspace = true
-name = "air_workspace"
-publish = false
-repository.workspace = true
-rust-version.workspace = true
-version = "0.0.0"
-
-[dependencies]
-air_r_parser = { workspace = true }
-air_r_formatter = { workspace = true }
-biome_console = { workspace = true }
-biome_diagnostics = { workspace = true }
-biome_formatter = { workspace = true }
-biome_parser = { workspace = true }
-biome_rowan = { workspace = true }
-thiserror = { workspace = true }
-
-[lints]
-workspace = true
